Fix time on DS1307 malfunction

+++ lib/tinystd/itoa.h
@@ -1,6 +1,6 @@
 #ifndef TEST1_ITOA_H
 #define TEST1_ITOA_H
 
-unsigned long itoas(int value, char *sp, int radix);
+unsigned long itoas(long value, char *sp, int radix);
 
 #endif //TEST1_ITOA_H
